Move the lxdprofile from common to uniter
Move lxdprofile api from common to uniter as it no longer is required
to be common between various components.
