+++ api/uniter/lxdprofile.go
@@ -1,7 +1,7 @@
 // Copyright 2018 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package common
+package uniter
 
 import (
 	"github.com/juju/errors"
